feat: Refactor MassTransit configuration.
Updated builder to detect is RabbitMQ has been deployed within the environment.

Wait until started to create message queues.
