+++ Company.Course/Company.Course.Api/EventBusRoutes.cs
@@ -5,6 +5,13 @@ public static class Services
 {
     public static void AddEventBusRoutes(this WebApplication app)
     {
+        var rabbitMqAvailable = Environment.GetEnvironmentVariable("RabbitMQAvailable");
+
+        if (!string.IsNullOrWhiteSpace(rabbitMqAvailable) && rabbitMqAvailable == "false")
+        {
+            return;
+        }
+
         app.MapPost("/queue/course/created/{id}/{name}", Created);
 
         static async Task<IResult> Created(IPublishEndpoint publishEndpoint, Guid id, string name)
