+++ Company.Course/Company.Course.Api/Program.cs
@@ -1,6 +1,5 @@
 using Company.Course.Models;
 using Company.Course.Services;
-using MassTransit;
 using Microsoft.EntityFrameworkCore;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -20,35 +19,7 @@ builder.Services.AddSwaggerGen(c =>
     c.SwaggerDoc("v1", new() { Title = "Company.Course.Api", Version = "v1" });
 });
 
-builder.Services.AddMassTransit(config =>
-{
-    config.AddBus(provider => Bus.Factory.CreateUsingRabbitMq(rabbitMqConfig =>
-    {
-        var rabbitMQServer = builder.Configuration.GetSection("RabbitMQConfiguration").GetSection("RabbitMQServer").Value;
-        var rabbitMQUsername = builder.Configuration.GetSection("RabbitMQConfiguration").GetSection("RabbitMQUsername").Value;
-        var rabbitMQPassword = builder.Configuration.GetSection("RabbitMQConfiguration").GetSection("RabbitMQPassword").Value;
-
-        if (string.IsNullOrWhiteSpace(rabbitMQServer))
-        {
-            rabbitMQServer = "localhost";
-        }
-
-        rabbitMqConfig.Host(new Uri($"rabbitmq://{rabbitMQServer}:5672"), h =>
-        {
-            if (!string.IsNullOrWhiteSpace(rabbitMQUsername))
-            {
-                h.Username(rabbitMQUsername);
-            }
-
-            if (!string.IsNullOrWhiteSpace(rabbitMQPassword))
-            {
-                h.Password(rabbitMQPassword);
-            }
-        });
-    }));
-});
-
-builder.Services.AddMassTransitHostedService();
+builder.AddMassTransitConfiguration();
 
 var app = builder.Build();
