+++ Company.Department/Company.Department.Api/EventBusRoutes.cs
@@ -5,6 +5,13 @@ public static class Services
 {
     public static void AddEventBusRoutes(this WebApplication app)
     {
+        var rabbitMqAvailable = Environment.GetEnvironmentVariable("RabbitMQAvailable");
+
+        if (!string.IsNullOrWhiteSpace(rabbitMqAvailable) && rabbitMqAvailable == "false")
+        {
+            return;
+        }
+
         app.MapPost("/queue/delete/department/{id}", Deleted);
 
         static async Task<IResult> Deleted(IPublishEndpoint publishEndpoint, Guid id)
