+++ Company.Employee/Company.Employee.Api/Program.cs
@@ -1,6 +1,5 @@
 using Company.Employee.Models;
 using Company.Employee.Services;
-using MassTransit;
 using Microsoft.EntityFrameworkCore;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -20,48 +19,7 @@ builder.Services.AddSwaggerGen(c =>
     c.SwaggerDoc("v1", new() { Title = "Company.Employee.Api", Version = "v1" });
 });
 
-builder.Services.AddMassTransit(config =>
-{
-    var rabbitMQServer = builder.Configuration.GetSection("RabbitMQConfiguration").GetSection("RabbitMQServer").Value;
-    var rabbitMQUsername = builder.Configuration.GetSection("RabbitMQConfiguration").GetSection("RabbitMQUsername").Value;
-    var rabbitMQPassword = builder.Configuration.GetSection("RabbitMQConfiguration").GetSection("RabbitMQPassword").Value;
-
-    if (string.IsNullOrWhiteSpace(rabbitMQServer))
-    {
-        rabbitMQServer = "localhost";
-    }
-
-    config.AddConsumer<EmployeeDepartmentDeletedService>();
-    config.AddConsumer<EmployeeNotificationListService>();
-
-    config.AddBus(provider => Bus.Factory.CreateUsingRabbitMq(rabbitMqConfig =>
-    {
-        rabbitMqConfig.Host(new Uri($"rabbitmq://{rabbitMQServer}:5672"), h =>
-        {
-            if (!string.IsNullOrWhiteSpace(rabbitMQUsername))
-            {
-                h.Username(rabbitMQUsername);
-            }
-
-            if (!string.IsNullOrWhiteSpace(rabbitMQPassword))
-            {
-                h.Password(rabbitMQPassword);
-            }
-        });
-
-        rabbitMqConfig.ReceiveEndpoint("department-deleted", e =>
-        {
-            e.Consumer(() => new EmployeeDepartmentDeletedService());
-        });
-
-        rabbitMqConfig.ReceiveEndpoint("notification-list", e =>
-        {
-            e.Consumer(() => new EmployeeNotificationListService());
-        });
-    }));
-});
-
-builder.Services.AddMassTransitHostedService();
+builder.AddMassTransitConfiguration();
 
 var app = builder.Build();
