+++ Company.Notification/Company.Notification.Api/EventBusRoutes.cs
@@ -5,6 +5,13 @@ public static class Services
 {
     public static void AddEventBusRoutes(this WebApplication app)
     {
+        var rabbitMqAvailable = Environment.GetEnvironmentVariable("RabbitMQAvailable");
+
+        if (!string.IsNullOrWhiteSpace(rabbitMqAvailable) && rabbitMqAvailable == "false")
+        {
+            return;
+        }
+
         app.MapPost("/queue/notification/list/{courseId}/{courseName}", NotificationList);
 
         static async Task<IResult> NotificationList(IPublishEndpoint publishEndpoint, Guid courseId, string courseName)
