+++ Company.Notification/Company.Notification.Api/Program.cs
@@ -1,6 +1,5 @@
 using Company.Notification.Models;
 using Company.Notification.Services;
-using MassTransit;
 using Microsoft.EntityFrameworkCore;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -20,54 +19,7 @@ builder.Services.AddSwaggerGen(c =>
     c.SwaggerDoc("v1", new() { Title = "Company.Notification.Api", Version = "v1" });
 });
 
-builder.Services.AddMassTransit(config =>
-{
-    var rabbitMQServer = builder.Configuration.GetSection("RabbitMQConfiguration").GetSection("RabbitMQServer").Value;
-    var rabbitMQUsername = builder.Configuration.GetSection("RabbitMQConfiguration").GetSection("RabbitMQUsername").Value;
-    var rabbitMQPassword = builder.Configuration.GetSection("RabbitMQConfiguration").GetSection("RabbitMQPassword").Value;
-
-    if (string.IsNullOrWhiteSpace(rabbitMQServer))
-    {
-        rabbitMQServer = "localhost";
-    }
-
-    config.AddConsumer<NotificationCourseCreatedService>();
-    config.AddConsumer<NotificationEmployeeCreatedService>();
-    config.AddConsumer<NotificationEmployeeCourseBroadcastService>();
-
-    config.AddBus(provider => Bus.Factory.CreateUsingRabbitMq(rabbitMqConfig =>
-    {
-        rabbitMqConfig.Host(new Uri($"rabbitmq://{rabbitMQServer}:5672"), h =>
-        {
-            if (!string.IsNullOrWhiteSpace(rabbitMQUsername))
-            {
-                h.Username(rabbitMQUsername);
-            }
-
-            if (!string.IsNullOrWhiteSpace(rabbitMQPassword))
-            {
-                h.Password(rabbitMQPassword);
-            }
-        });
-
-        rabbitMqConfig.ReceiveEndpoint("course-created", e =>
-        {
-            e.Consumer(() => new NotificationCourseCreatedService());
-        });
-
-        rabbitMqConfig.ReceiveEndpoint("employee-created", e =>
-        {
-            e.Consumer(() => new NotificationEmployeeCreatedService());
-        });
-
-        rabbitMqConfig.ReceiveEndpoint("employee-course-broadcast", e =>
-        {
-            e.Consumer(() => new NotificationEmployeeCourseBroadcastService());
-        });
-    }));
-});
-
-builder.Services.AddMassTransitHostedService();
+builder.AddMassTransitConfiguration();
 
 var app = builder.Build();
