+++ Company.Notification/Company.Notification.Api/Properties/launchSettings.json
@@ -13,7 +13,8 @@
       "launchBrowser": true,
       "launchUrl": "swagger",
       "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development"
+        "ASPNETCORE_ENVIRONMENT": "Development",
+        "RabbitMQAvailable": "false"
       },
       "applicationUrl": "https://localhost:5004",
       "dotnetRunMessages": true
@@ -30,7 +31,11 @@
       "launchBrowser": true,
       "launchUrl": "{Scheme}://{ServiceHost}:{ServicePort}",
       "publishAllPorts": true,
-      "useSSL": true
+      "useSSL": true,
+      "environmentVariables": {
+        "ASPNETCORE_ENVIRONMENT": "Development",
+        "RabbitMQAvailable": "false"
+      }
     }
   }
 }
